Fix member expiration not always working
The Sidekiq job `RemoveExpiredMembersWorker` was failing to run in
production because it was hitting statement timeouts because it was
scanning all rows in order. On staging, where it used to scan 4 million
rows, adding an index brought this down to only a few hundred rows.

This commit backports the fix for 12-3-auto-deploy-20190911.

See https://gitlab.com/gitlab-org/gitlab-ce/issues/67286