Skip gitlab:graphql:check_docs task unless EE.
The rake task is being run in a pipeline job on gitlab-foss. It will
always break on gitlab-foss, as it checks whether the docs need to be
updated by comparing the GraphQL Schema against the documentation, which
will be different in gitlab-foss because it’s missing all the
EE-prepended schema.

https://gitlab.com/gitlab-org/gitlab/issues/34293