Validate Issue#closed_at after #created_at
A very small number of Issue records have a closed_at time that is
after their created_at time. Although we suspect this is related to
import, it is difficult to troubleshoot and we have not yet isolated
the bug. This validation will help isolate the bug, as it will only
fail in the case when the bug has manifested. This validation is
wrapped in a feature flag called validate_issues_closed_after_created
that defaults to being enabled, so it can be disabled by users who
encounter this bug and want to import anyway.