Only use Elasticsearch if all projects have been indexed
Since we'll be rolling out Elasticsearch gradually we want to ensure
that users don't start using the new Elasticsearch until we've actually
indexed all projects within the group.

If all projects in the group have been indexed once then we can search
with Elasticsearch.When searching in the context of a single project
then we also check that this project has been indexed before using
Elasticsearch so it is not just applying to groups.

NOTE: For now this does not query to confirm that the repos have
finished indexing. This would be somewhat tricky because we could look
at `index_statuses.last_commit` but this is never set for projects that
have no repository and same goes for wiki.