Add ability to duplicate system dashboards
Adds a new item to the dashboard dropdown to allow the user
to duplicate the dashboard into their projects.

Add a modal to create a new dashboard based on a system dashboard,
the user can provide file name and commit message for the new file.

Refactors the dropdown into a separate component to reduce
the dashboard complexity.