Add index services on template
This migration is a corrective action to add the missing
index_services_on_template index on staging.

We suspect the column rename was executed in staging already before it
was rolled back/reverted with 8d4d9705. The rollback doesn't know about
the existence of the index though and hence it is now missing.