Add button to return to MR on pipeline celebration
The pipeline suggest 1st nudge starts on the merge request, so:

- the 1st nudge was changed to pass the merge request path from
the 1st to the 2nd/3rd nudge.

- the 3rd nudge was changed to set the Merge request path into
the cookie set when clicking the Commit Changes button.

- the 4th nudge then checks for the cookie name instead of just
the cookie existence. The merge request path is used for the new
button. If they happened to have an old cookie, it links to the
project merge request list instead.