Add FormInput component
This adds a form input component, to be used in the [SAST Configuration
UI][epic].

This component is a fairly thin wrapper around the `GlFormGroup` and
`GlFormInput` components, to conform to the [design]. It will be
consumed by a [parent container component][container], itself driven via
a schema to generate a form.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/225224.

[epic]: https://gitlab.com/groups/gitlab-org/-/epics/3659
[container]: https://gitlab.com/gitlab-org/gitlab/-/issues/231370
[design]: https://gitlab.com/gitlab-org/gitlab/-/issues/228830