Fix overflow issues with monaco file editor
*What happened?*
`pre` has some `overflow: auto` rules which
causes issues when loading the monaco
editor that expects to not have `overflow: visible`
on the parent.

*What's the fix?*
We really shouldn't put elements under a `pre`, so
this MR moves `pre` inside the #editor elements
to help present the content while the editor
bit is loading.